+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -284,10 +284,12 @@ instance GenValid NonNegativeInterval where
   shrinkValid _ = []
 
 instance Arbitrary TxIx where
-  arbitrary = TxIx <$> arbitrary
+  -- starting with Conway, we only deserialize TxIx within Word16 range
+  arbitrary = TxIx . fromIntegral <$> arbitrary @Word16
 
 instance Arbitrary CertIx where
-  arbitrary = CertIx <$> arbitrary
+  -- starting with Conway, we only deserialize CertIx within Word16 range
+  arbitrary = CertIx . fromIntegral <$> arbitrary @Word16
 
 instance Arbitrary ProtVer where
   arbitrary = ProtVer <$> arbitrary <*> arbitrary